components/deployment/requirements.txt
google-cloud-aiplatform==1.50.0
